requirements.txt
# Core MCP and HTTP dependencies
mcp==1.2.1
httpx==0.28.1
click==8.1.8
anyio==4.8.0

# OpenMetadata client
openmetadata-ingestion==1.6.1

# Configuration and utilities
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Core dependencies
annotated-types==0.7.0
certifi==2025.1.31
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx-sse==0.4.0
idna==3.10
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.41.3
typing-extensions==4.12.2

# Remote server capabilities (optional)
fastapi==0.115.5
uvicorn==0.32.1
websockets==14.1

# Authentication features (optional)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
cryptography==44.0.0
bcrypt==4.2.1

# Google OAuth (optional)
google-auth>=2.40.3
google-auth-oauthlib>=1.2.2
google-auth-httplib2>=0.2.0

# Monitoring and logging (optional)
sentry-sdk[fastapi]==2.19.2
structlog==25.1.0

# Web dashboard (optional)
jinja2==3.1.5
aiofiles==24.1.0
